rename module crate::command::{describe,init}
